fix: wip

+++ packages/lib/server-only/public-api/delete-api-token-by-id.ts
@@ -1,32 +1,30 @@
-import { TeamMemberRole } from '@prisma/client';
-
 import { prisma } from '@documenso/prisma';
+
+import { TEAM_MEMBER_ROLE_PERMISSIONS_MAP } from '../../constants/teams';
+import { AppError, AppErrorCode } from '../../errors/app-error';
+import { buildTeamWhereQuery } from '../../utils/teams';
 
 export type DeleteTokenByIdOptions = {
   id: number;
   userId: number;
-  teamId?: number;
+  teamId: number;
 };
 
 export const deleteTokenById = async ({ id, userId, teamId }: DeleteTokenByIdOptions) => {
-  if (teamId) {
-    const member = await prisma.teamMember.findFirst({
-      where: {
-        userId,
-        teamId,
-        role: TeamMemberRole.ADMIN,
-      },
-    });
+  const team = await prisma.team.findFirst({
+    where: buildTeamWhereQuery(teamId, userId, TEAM_MEMBER_ROLE_PERMISSIONS_MAP['MANAGE_TEAM']),
+  });
 
-    if (!member) {
-      throw new Error('You do not have permission to delete this token');
-    }
+  if (!team) {
+    throw new AppError(AppErrorCode.UNAUTHORIZED, {
+      message: 'You do not have permission to delete this token',
+    });
   }
 
   return await prisma.apiToken.delete({
     where: {
       id,
-      teamId: teamId ?? null,
+      teamId,
     },
   });
 };
